agora-api/pyproject.toml
[tool.poetry]
name = "agora-api"
version = "0.1.0"
description = ""
authors = ["Will Meyburgh <annekinmeyburgh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
google-genai = "^1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
